Move pluralizers into common code
