+++ services/app-web/src/common-code/formatters/index.ts
@@ -1,0 +1,3 @@
+import { pluralize } from './pluralizer'
+
+export { pluralize }
